+++ packages/aspida-swagger/samples/jsonV2/pet/index.ts
@@ -1,8 +1,0 @@
-/* eslint-disable */
-export interface Methods {
-  post: {
-  }
-
-  put: {
-  }
-}