+++ packages/aspida-swagger/samples/jsonV2/store/order/_orderId@number.ts
@@ -1,8 +1,0 @@
-/* eslint-disable */
-export interface Methods {
-  get: {
-  }
-
-  delete: {
-  }
-}