+++ packages/aspida-swagger/samples/yamlV2/user/logout.ts
@@ -1,5 +1,0 @@
-/* eslint-disable */
-export interface Methods {
-  get: {
-  }
-}