+++ packages/aspida/src/getConfig.ts
@@ -4,14 +4,16 @@ import path from 'path'
 export interface BaseConfig {
   input: string
   baseURL: string
+  trailingSlash: boolean
 }
 
 interface ConfigFile {
   input: string
   baseURL?: string
+  trailingSlash?: boolean
 }
 
-const defaultConfig = { input: 'apis', baseURL: '' }
+const defaultConfig = { input: 'apis', baseURL: '', trailingSlash: false }
 
 export default (configPath = 'aspida.config.js'): BaseConfig[] => {
   if (fs.existsSync(configPath)) {