+++ packages/openapi2aspida/samples/swagger/pet/index.ts
@@ -1,0 +1,12 @@
+/* eslint-disable */
+import * as Types from '../@types'
+
+export interface Methods {
+  post: {
+    reqData: Types.Pet
+  }
+
+  put: {
+    reqData: Types.Pet
+  }
+}