+++ packages/openapi2aspida/samples/swagger/user/login.ts
@@ -1,0 +1,16 @@
+/* eslint-disable */
+export interface Methods {
+  get: {
+    query: {
+      username: string
+      password: string
+    }
+
+    resData: string
+
+    resHeaders: {
+      'X-Expires-After': string
+      'X-Rate-Limit': number
+    }
+  }
+}