+++ packages/pathpida/samples/index.tsx
@@ -1,0 +1,7 @@
+
+
+export interface Query {
+  hoge: number
+  fuga: string[]
+}
+